sdm710-common: decommonize sensors blobs
Sensors HAL are tightly connected with camera HAL. Due to facts that
sirius probably will not receive A11 update and camera HAL is highly
device specific we cannot update sensors HAL in sdm710-common.
We have to decommonize sensors HAL to be able to update camera HAL
on other devices.

Change-Id: If1f91c46609d909e9c49c44e31b6fd9adb20f1a8
Signed-off-by: Ivan Vecera <ivan@cera.cz>
Signed-off-by: DarkAngelGR <asavvo01@gmail.com>